[front] checkout: move prepare-payment logic to lib/api/checkout
