+++ front/pages/api/w/[wId]/subscriptions/checkout/prepare-payment.ts
@@ -1,34 +1,20 @@
 // @migration-status: MIGRATED_TO_HONO
 /** @ignoreswagger */
+
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
-import { getStripeCheckoutSessionStatus } from "@app/lib/api/stripe/checkout_status";
-import { isMetronomeBillingEnabled } from "@app/lib/api/subscription";
+import {
+  type GetPreparePaymentResponseBody,
+  getPreparePaymentData,
+  type PreparePaymentError,
+} from "@app/lib/api/checkout/prepare_payment";
 import type { Authenticator } from "@app/lib/auth";
-import { getBillingCurrencyForCountry } from "@app/lib/plans/billing_currency";
-import { calculateTax, getStripeClient } from "@app/lib/plans/stripe";
-import { CouponResource } from "@app/lib/resources/coupon_resource";
 import { apiError } from "@app/logger/withlogging";
-import type { SupportedCurrency } from "@app/types/currency";
 import type { WithAPIErrorResponse } from "@app/types/error";
+import { assertNever } from "@app/types/shared/utils/assert_never";
 import { isString } from "@app/types/shared/utils/general";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-export type GetPreparePaymentResponseBody =
-  | { status: "pending" }
-  | {
-      status: "success";
-      subtotalCents: number;
-      taxCents: number;
-      totalCents: number;
-      seatCount: number;
-      pricePerSeatCents: number;
-      planCode: string;
-      metronomePackageAlias: string;
-      currency: SupportedCurrency;
-      cardBrand?: string;
-      cardLast4?: string;
-      sepaLast4?: string;
-    };
+export type { GetPreparePaymentResponseBody };
 
 async function handler(
   req: NextApiRequest,
@@ -55,21 +41,6 @@ async function handler(
       },
     });
   }
-  const owner = auth.getNonNullableWorkspace();
-
-  const useMetronomeBilling = await isMetronomeBillingEnabled(auth);
-  if (!useMetronomeBilling) {
-    return apiError(req, res, {
-      status_code: 403,
-      api_error: {
-        type: "workspace_auth_error",
-        message: "Metronome billing is not enabled for this workspace.",
-      },
-    });
-  }
-
-  // Prevent HTTP caching as session status can change on every call.
-  res.setHeader("Cache-Control", "no-store");
 
   const { setup_session_id } = req.query;
   if (!isString(setup_session_id)) {
@@ -82,145 +53,83 @@ async function handler(
     });
   }
 
-  // onComplete fires on the client before Stripe marks the session complete server-side.
-  // We return "pending" so the client can retry instead of blocking here.
-  const sessionStatus = await getStripeCheckoutSessionStatus(setup_session_id);
-  if (sessionStatus?.status !== "complete") {
-    return res.status(200).json({ status: "pending" });
+  // Prevent HTTP caching as session status can change on every call.
+  res.setHeader("Cache-Control", "no-store");
+
+  const result = await getPreparePaymentData(auth, setup_session_id);
+  if (result.isErr()) {
+    return mapPreparePaymentError(req, res, result.error);
   }
 
-  const stripe = getStripeClient();
-  const setupSession = await stripe.checkout.sessions.retrieve(
-    setup_session_id,
-    { expand: ["setup_intent", "setup_intent.payment_method"] }
-  );
+  return res.status(200).json(result.value);
+}
 
-  const setupIntent = setupSession.setup_intent;
-  if (
-    !setupIntent ||
-    isString(setupIntent) ||
-    setupIntent.status !== "succeeded"
-  ) {
-    return apiError(req, res, {
-      status_code: 400,
-      api_error: {
-        type: "invalid_request_error",
-        message: "Setup session has not completed successfully.",
-      },
-    });
+function mapPreparePaymentError(
+  req: NextApiRequest,
+  res: NextApiResponse<WithAPIErrorResponse<GetPreparePaymentResponseBody>>,
+  error: PreparePaymentError
+): void {
+  switch (error.type) {
+    case "metronome_not_enabled":
+      return apiError(req, res, {
+        status_code: 403,
+        api_error: {
+          type: "workspace_auth_error",
+          message: "Metronome billing is not enabled for this workspace.",
+        },
+      });
+    case "setup_not_succeeded":
+      return apiError(req, res, {
+        status_code: 400,
+        api_error: {
+          type: "invalid_request_error",
+          message: "Setup session has not completed successfully.",
+        },
+      });
+    case "workspace_mismatch":
+      return apiError(req, res, {
+        status_code: 403,
+        api_error: {
+          type: "workspace_auth_error",
+          message: "Setup intent does not correspond to the current workspace.",
+        },
+      });
+    case "missing_metadata":
+      return apiError(req, res, {
+        status_code: 500,
+        api_error: {
+          type: "internal_server_error",
+          message:
+            "Setup session metadata is missing seatCount or pricePerSeatCents.",
+        },
+      });
+    case "missing_customer_id":
+      return apiError(req, res, {
+        status_code: 500,
+        api_error: {
+          type: "internal_server_error",
+          message: "Setup session is missing a customer ID.",
+        },
+      });
+    case "customer_deleted":
+      return apiError(req, res, {
+        status_code: 500,
+        api_error: {
+          type: "internal_server_error",
+          message: "Stripe customer has been deleted.",
+        },
+      });
+    case "tax_calculation_failed":
+      return apiError(req, res, {
+        status_code: 500,
+        api_error: {
+          type: "internal_server_error",
+          message: error.message,
+        },
+      });
+    default:
+      assertNever(error);
   }
-  if (setupSession.client_reference_id !== owner.sId) {
-    return apiError(req, res, {
-      status_code: 403,
-      api_error: {
-        type: "workspace_auth_error",
-        message: "Setup intent does not correspond to the current workspace.",
-      },
-    });
-  }
-
-  const { seatCount: seatCountStr, pricePerSeatCents: pricePerSeatCentsStr } =
-    setupSession.metadata ?? {};
-
-  if (!isString(seatCountStr) || !isString(pricePerSeatCentsStr)) {
-    return apiError(req, res, {
-      status_code: 500,
-      api_error: {
-        type: "internal_server_error",
-        message:
-          "Setup session metadata is missing seatCount or pricePerSeatCents.",
-      },
-    });
-  }
-
-  const seatCount = Number(seatCountStr);
-  const pricePerSeatCents = Number(pricePerSeatCentsStr);
-  const subtotalCents = seatCount * pricePerSeatCents;
-  const stripeCustomerId = setupSession.customer;
-
-  if (!isString(stripeCustomerId)) {
-    return apiError(req, res, {
-      status_code: 500,
-      api_error: {
-        type: "internal_server_error",
-        message: "Setup session is missing a customer ID.",
-      },
-    });
-  }
-
-  const planCode = setupSession.metadata?.planCode ?? "";
-  const metronomePackageAlias =
-    setupSession.metadata?.metronomePackageAlias ?? "";
-
-  const rawPaymentMethod = setupIntent.payment_method;
-  let cardBrand: string | undefined;
-  let cardLast4: string | undefined;
-  let sepaLast4: string | undefined;
-  if (rawPaymentMethod && !isString(rawPaymentMethod)) {
-    if (rawPaymentMethod.card) {
-      cardBrand = rawPaymentMethod.card.brand;
-      cardLast4 = rawPaymentMethod.card.last4;
-    } else if (rawPaymentMethod.sepa_debit) {
-      sepaLast4 = rawPaymentMethod.sepa_debit.last4 ?? undefined;
-    }
-  }
-
-  const customer = await stripe.customers.retrieve(stripeCustomerId);
-  if (customer.deleted) {
-    return apiError(req, res, {
-      status_code: 500,
-      api_error: {
-        type: "internal_server_error",
-        message: "Stripe customer has been deleted.",
-      },
-    });
-  }
-  const country = customer.address?.country ?? "US";
-  const currency = getBillingCurrencyForCountry(country, true);
-
-  // Apply coupon discount to the tax base if a coupon was stored in session metadata.
-  // No validation here — enforcement happens in POST /payment.
-  const couponCode = setupSession.metadata?.couponCode;
-  let discountedSubtotalCents = subtotalCents;
-  if (couponCode) {
-    const coupon = await CouponResource.findByCode(couponCode);
-    if (coupon) {
-      discountedSubtotalCents = Math.max(
-        0,
-        subtotalCents - coupon.amount * 100
-      );
-    }
-  }
-
-  const taxResult = await calculateTax({
-    stripeCustomerId,
-    amountCents: discountedSubtotalCents,
-    currency,
-  });
-  if (taxResult.isErr()) {
-    return apiError(req, res, {
-      status_code: 500,
-      api_error: {
-        type: "internal_server_error",
-        message: taxResult.error.error_message,
-      },
-    });
-  }
-
-  return res.status(200).json({
-    status: "success",
-    subtotalCents,
-    taxCents: taxResult.value.taxCents,
-    totalCents: taxResult.value.totalCents,
-    seatCount,
-    pricePerSeatCents,
-    planCode,
-    metronomePackageAlias,
-    currency,
-    cardBrand,
-    cardLast4,
-    sepaLast4,
-  });
 }
 
 export default withSessionAuthenticationForWorkspace(handler, {
